Change strategy of output.
Use a different strategy to retrieve the content of STDOUT and STDERR.
The `ShellOutError` structure has been changed but the API remains the same.
The structure now contains both STDOUT and STDERR as `Data` and we can
retrieve the equivalent `String` using the previous properties `message`
and `output`.
From my memories, any non 0 `terminationStatus` should be considered as
an error.
This also allows to introduce a parameter `printOutput` which will
output the received data to `STDOUT` and `STDERR`.