Enable a custom Process instance to be injected
This change enables a custom `Process` instance to be injected when
using `shellOut`. This in turns enables call sites to monitor and
terminate the processes they start, which is useful for long-running
commands that may need to be terminated at some point.

(Also includes a minor git ignore update for modern versions of SwiftPM,
as well as some code cleanup)