Updated to Swift 5.1 and fixed deprecations
Remove "#if os" checks, since `readabilityHandler` was implemented on Linux in Swift 5.1.
Added "#if available" checks to use `executableURL` and `run()` instead of `launchPath` and `launch()` on macOS 10.13+ and Linux.
Added missing tests to `allTests` so they run on Linux.
Changed `testSingleCommandAtPathContainingTilde()` as it formerly failed with an empty home directory (common on Linux and Docker images).